+++ src/app.module.ts
@@ -6,9 +6,18 @@ import { InvoicesModule } from "./modules/invoices.module";
 import { ExpensesModule } from "./modules/expenses.module";
 import { FixedExpensesModule } from "./modules/fixed-expenses.module";
 import { FreelanceInvoiceModule } from "./modules/freelance-invoice.module";
+import { CacheModule } from "./modules/cache.module";
 
 @Module({
-  imports: [AuthModule, UsersModule, InvoicesModule, ExpensesModule, FixedExpensesModule, FreelanceInvoiceModule],
+  imports: [
+    CacheModule,
+    AuthModule,
+    UsersModule,
+    InvoicesModule,
+    ExpensesModule,
+    FixedExpensesModule,
+    FreelanceInvoiceModule,
+  ],
   controllers: [HealthController],
 })
 export class AppModule {}